Extracted GENNExampleApp file
